Re-structure mouse IO
This commit re-structures the way mouse IO is handled and consequently how
forces and densities are updated.

Forces are now generated through drag-and-release motions, meaning the force is
released when the mouse button is released. The force is determined by the x and
y components of the vector generated by the drag-and-release motion. The force
is released at the point of which the mouse was pressed.

Generating densities works just like before, the only thing that has changed is
the structure of how this is done in the code.

See diff for changes in code structure regarding mouse IO.